requirements.txt
streamlit==1.26.0
opencv-python-headless==4.7.0.72
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
Pillow==9.5.0